# Rewrite context watcher to use single background goroutine