@jackc 2022-09-24
I have a few concerns about this change.

I'm not sure atomics as used provide the same synchronization that the mutex did. Maybe it does... But it is definitely harder to reason about. My experience with atomics has led me to avoid them unless there is a very large performance gain. I'd expect that this is a low contention lock where the mutex would be pretty fast.

But regarding the persistent background goroutine overall -- have you benchmarked how this would affect entire queries? I went down this path myself several years ago and found at best no significant improvements. IIRC, the additional synchronization cost offset the savings from starting a goroutine. Maybe things have changed since then but unless there are major benefits I'd rather stay with the simpler approach we have now.